--- raft/node_test.go
@@ -1,805 +1,0 @@
-// Copyright 2015 The etcd Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package raft
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"math"
-	"reflect"
-	"strings"
-	"testing"
-	"time"
-
-	pb "github.com/pingcap-incubator/tinykv/proto/pkg/eraftpb"
-)
-
-// readyWithTimeout selects from n.Ready() with a 1-second timeout. It
-// panics on timeout, which is better than the indefinite wait that
-// would occur if this channel were read without being wrapped in a
-// select.
-func readyWithTimeout(n Node) Ready {
-	select {
-	case rd := <-n.Ready():
-		return rd
-	case <-time.After(time.Second):
-		panic("timed out waiting for ready")
-	}
-}
-
-// TestNodeStep ensures that node.Step sends MessageType_MsgPropose to propc chan
-// and other kinds of messages to recvc chan.
-func TestNodeStep(t *testing.T) {
-	for i, msgn := range pb.MessageType_name {
-		n := &node{
-			propc: make(chan msgWithResult, 1),
-			recvc: make(chan pb.Message, 1),
-		}
-		msgt := pb.MessageType(i)
-		n.Step(context.TODO(), pb.Message{MsgType: msgt})
-		// Proposal goes to proc chan. Others go to recvc chan.
-		if msgt == pb.MessageType_MsgPropose {
-			select {
-			case <-n.propc:
-			default:
-				t.Errorf("%d: cannot receive %s on propc chan", msgt, msgn)
-			}
-		} else {
-			if IsLocalMsg(msgt) {
-				select {
-				case <-n.recvc:
-					t.Errorf("%d: step should ignore %s", msgt, msgn)
-				default:
-				}
-			} else {
-				select {
-				case <-n.recvc:
-				default:
-					t.Errorf("%d: cannot receive %s on recvc chan", msgt, msgn)
-				}
-			}
-		}
-	}
-}
-
-// Cancel and Stop should unblock Step()
-func TestNodeStepUnblock(t *testing.T) {
-	// a node without buffer to block step
-	n := &node{
-		propc: make(chan msgWithResult),
-		done:  make(chan struct{}),
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	stopFunc := func() { close(n.done) }
-
-	tests := []struct {
-		unblock func()
-		werr    error
-	}{
-		{stopFunc, ErrStopped},
-		{cancel, context.Canceled},
-	}
-
-	for i, tt := range tests {
-		errc := make(chan error, 1)
-		go func() {
-			err := n.Step(ctx, pb.Message{MsgType: pb.MessageType_MsgPropose})
-			errc <- err
-		}()
-		tt.unblock()
-		select {
-		case err := <-errc:
-			if err != tt.werr {
-				t.Errorf("#%d: err = %v, want %v", i, err, tt.werr)
-			}
-			//clean up side-effect
-			if ctx.Err() != nil {
-				ctx = context.TODO()
-			}
-			select {
-			case <-n.done:
-				n.done = make(chan struct{})
-			default:
-			}
-		case <-time.After(1 * time.Second):
-			t.Fatalf("#%d: failed to unblock step", i)
-		}
-	}
-}
-
-// TestNodePropose ensures that node.Propose sends the given proposal to the underlying raft.
-func TestNodePropose(t *testing.T) {
-	msgs := []pb.Message{}
-	appendStep := func(r *Raft, m pb.Message) error {
-		msgs = append(msgs, m)
-		return nil
-	}
-
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	go n.run(r)
-	n.Campaign(context.TODO())
-	for {
-		rd := <-n.Ready()
-		s.Append(rd.Entries)
-		// change the step function to appendStep until this raft becomes leader
-		if rd.SoftState.Lead == r.id {
-			r.step = appendStep
-			n.Advance()
-			break
-		}
-		n.Advance()
-	}
-	n.Propose(context.TODO(), []byte("somedata"))
-	n.Stop()
-
-	if len(msgs) != 1 {
-		t.Fatalf("len(msgs) = %d, want %d", len(msgs), 1)
-	}
-	if msgs[0].MsgType != pb.MessageType_MsgPropose {
-		t.Errorf("msg type = %d, want %d", msgs[0].MsgType, pb.MessageType_MsgPropose)
-	}
-	if !bytes.Equal(msgs[0].Entries[0].Data, []byte("somedata")) {
-		t.Errorf("data = %v, want %v", msgs[0].Entries[0].Data, []byte("somedata"))
-	}
-}
-
-// TestNodeProposeConfig ensures that node.ProposeConfChange sends the given configuration proposal
-// to the underlying raft.
-func TestNodeProposeConfig(t *testing.T) {
-	msgs := []pb.Message{}
-	appendStep := func(r *Raft, m pb.Message) error {
-		msgs = append(msgs, m)
-		return nil
-	}
-
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	go n.run(r)
-	n.Campaign(context.TODO())
-	for {
-		rd := <-n.Ready()
-		s.Append(rd.Entries)
-		// change the step function to appendStep until this raft becomes leader
-		if rd.SoftState.Lead == r.id {
-			r.step = appendStep
-			n.Advance()
-			break
-		}
-		n.Advance()
-	}
-	cc := pb.ConfChange{ChangeType: pb.ConfChangeType_AddNode, NodeId: 1}
-	ccdata, err := cc.Marshal()
-	if err != nil {
-		t.Fatal(err)
-	}
-	n.ProposeConfChange(context.TODO(), cc)
-	n.Stop()
-
-	if len(msgs) != 1 {
-		t.Fatalf("len(msgs) = %d, want %d", len(msgs), 1)
-	}
-	if msgs[0].MsgType != pb.MessageType_MsgPropose {
-		t.Errorf("msg type = %d, want %d", msgs[0].MsgType, pb.MessageType_MsgPropose)
-	}
-	if !bytes.Equal(msgs[0].Entries[0].Data, ccdata) {
-		t.Errorf("data = %v, want %v", msgs[0].Entries[0].Data, ccdata)
-	}
-}
-
-// TestNodeProposeAddDuplicateNode ensures that two proposes to add the same node should
-// not affect the later propose to add new node.
-func TestNodeProposeAddDuplicateNode(t *testing.T) {
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	go n.run(r)
-	n.Campaign(context.TODO())
-	rdyEntries := make([]*pb.Entry, 0)
-	ticker := time.NewTicker(time.Millisecond * 100)
-	defer ticker.Stop()
-	done := make(chan struct{})
-	stop := make(chan struct{})
-	applyConfChan := make(chan struct{})
-
-	go func() {
-		defer close(done)
-		for {
-			select {
-			case <-stop:
-				return
-			case <-ticker.C:
-				n.Tick()
-			case rd := <-n.Ready():
-				s.Append(rd.Entries)
-				applied := false
-				for _, e := range rd.Entries {
-					rdyEntries = append(rdyEntries, &e)
-					switch e.EntryType {
-					case pb.EntryType_EntryNormal:
-					case pb.EntryType_EntryConfChange:
-						var cc pb.ConfChange
-						cc.Unmarshal(e.Data)
-						n.ApplyConfChange(cc)
-						applied = true
-					}
-				}
-				n.Advance()
-				if applied {
-					applyConfChan <- struct{}{}
-				}
-			}
-		}
-	}()
-
-	cc1 := pb.ConfChange{ChangeType: pb.ConfChangeType_AddNode, NodeId: 1}
-	ccdata1, _ := cc1.Marshal()
-	n.ProposeConfChange(context.TODO(), cc1)
-	<-applyConfChan
-
-	// try add the same node again
-	n.ProposeConfChange(context.TODO(), cc1)
-	<-applyConfChan
-
-	// the new node join should be ok
-	cc2 := pb.ConfChange{ChangeType: pb.ConfChangeType_AddNode, NodeId: 2}
-	ccdata2, _ := cc2.Marshal()
-	n.ProposeConfChange(context.TODO(), cc2)
-	<-applyConfChan
-
-	close(stop)
-	<-done
-
-	if len(rdyEntries) != 4 {
-		t.Errorf("len(entry) = %d, want %d, %v\n", len(rdyEntries), 4, rdyEntries)
-	}
-	if !bytes.Equal(rdyEntries[1].Data, ccdata1) {
-		t.Errorf("data = %v, want %v", rdyEntries[1].Data, ccdata1)
-	}
-	if !bytes.Equal(rdyEntries[3].Data, ccdata2) {
-		t.Errorf("data = %v, want %v", rdyEntries[3].Data, ccdata2)
-	}
-	n.Stop()
-}
-
-// TestBlockProposal ensures that node will block proposal when it does not
-// know who is the current leader; node will accept proposal when it knows
-// who is the current leader.
-func TestBlockProposal(t *testing.T) {
-	n := newNode()
-	r := newTestRaft(1, []uint64{1}, 10, 1, NewMemoryStorage())
-	go n.run(r)
-	defer n.Stop()
-
-	errc := make(chan error, 1)
-	go func() {
-		errc <- n.Propose(context.TODO(), []byte("somedata"))
-	}()
-
-	// briefly sleeps in order to invoke the go scheduler.
-	time.Sleep(10 * time.Millisecond)
-	select {
-	case err := <-errc:
-		t.Errorf("err = %v, want blocking", err)
-	default:
-	}
-
-	n.Campaign(context.TODO())
-	select {
-	case err := <-errc:
-		if err != nil {
-			t.Errorf("err = %v, want %v", err, nil)
-		}
-	case <-time.After(10 * time.Second):
-		t.Errorf("blocking proposal, want unblocking")
-	}
-}
-
-func TestNodeProposeWaitDropped(t *testing.T) {
-	msgs := []pb.Message{}
-	droppingMsg := []byte("test_dropping")
-	dropStep := func(r *Raft, m pb.Message) error {
-		if m.MsgType == pb.MessageType_MsgPropose && strings.Contains(m.String(), string(droppingMsg)) {
-			t.Logf("dropping message: %v", m.String())
-			return ErrProposalDropped
-		}
-		msgs = append(msgs, m)
-		return nil
-	}
-
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	go n.run(r)
-	n.Campaign(context.TODO())
-	for {
-		rd := <-n.Ready()
-		s.Append(rd.Entries)
-		// change the step function to dropStep until this raft becomes leader
-		if rd.SoftState.Lead == r.id {
-			r.step = dropStep
-			n.Advance()
-			break
-		}
-		n.Advance()
-	}
-	proposalTimeout := time.Millisecond * 100
-	ctx, cancel := context.WithTimeout(context.Background(), proposalTimeout)
-	// propose with cancel should be cancelled earyly if dropped
-	err := n.Propose(ctx, droppingMsg)
-	if err != ErrProposalDropped {
-		t.Errorf("should drop proposal : %v", err)
-	}
-	cancel()
-
-	n.Stop()
-	if len(msgs) != 0 {
-		t.Fatalf("len(msgs) = %d, want %d", len(msgs), 1)
-	}
-}
-
-// TestNodeTick ensures that node.Tick() will increase the
-// elapsed of the underlying raft state machine.
-func TestNodeTick(t *testing.T) {
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	go n.run(r)
-	elapsed := r.electionElapsed
-	n.Tick()
-
-	for len(n.tickc) != 0 {
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	n.Stop()
-	if r.electionElapsed != elapsed+1 {
-		t.Errorf("elapsed = %d, want %d", r.electionElapsed, elapsed+1)
-	}
-}
-
-// TestNodeStop ensures that node.Stop() blocks until the node has stopped
-// processing, and that it is idempotent
-func TestNodeStop(t *testing.T) {
-	n := newNode()
-	s := NewMemoryStorage()
-	r := newTestRaft(1, []uint64{1}, 10, 1, s)
-	donec := make(chan struct{})
-
-	go func() {
-		n.run(r)
-		close(donec)
-	}()
-
-	status := n.Status()
-	n.Stop()
-
-	select {
-	case <-donec:
-	case <-time.After(time.Second):
-		t.Fatalf("timed out waiting for node to stop!")
-	}
-
-	emptyStatus := Status{}
-
-	if reflect.DeepEqual(status, emptyStatus) {
-		t.Errorf("status = %v, want not empty", status)
-	}
-	// Further status should return be empty, the node is stopped.
-	status = n.Status()
-	if !reflect.DeepEqual(status, emptyStatus) {
-		t.Errorf("status = %v, want empty", status)
-	}
-	// Subsequent Stops should have no effect.
-	n.Stop()
-}
-
-func TestReadyContainUpdates(t *testing.T) {
-	tests := []struct {
-		rd       Ready
-		wcontain bool
-	}{
-		{Ready{}, false},
-		{Ready{SoftState: &SoftState{Lead: 1}}, true},
-		{Ready{HardState: pb.HardState{Vote: 1}}, true},
-		{Ready{Entries: make([]pb.Entry, 1)}, true},
-		{Ready{CommittedEntries: make([]pb.Entry, 1)}, true},
-		{Ready{Messages: make([]pb.Message, 1)}, true},
-		{Ready{Snapshot: pb.Snapshot{Metadata: &pb.SnapshotMetadata{Index: 1}}}, true},
-	}
-
-	for i, tt := range tests {
-		if g := tt.rd.containsUpdates(); g != tt.wcontain {
-			t.Errorf("#%d: containUpdates = %v, want %v", i, g, tt.wcontain)
-		}
-	}
-}
-
-// TestNodeStart ensures that a node can be started correctly. The node should
-// start with correct configuration change entries, and can accept and commit
-// proposals.
-func TestNodeStart(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	cc := pb.ConfChange{ChangeType: pb.ConfChangeType_AddNode, NodeId: 1}
-	ccdata, err := cc.Marshal()
-	if err != nil {
-		t.Fatalf("unexpected marshal error: %v", err)
-	}
-	wants := []Ready{
-		{
-			HardState: pb.HardState{Term: 1, Commit: 1, Vote: 0},
-			Entries: []pb.Entry{
-				{EntryType: pb.EntryType_EntryConfChange, Term: 1, Index: 1, Data: ccdata},
-			},
-			CommittedEntries: []pb.Entry{
-				{EntryType: pb.EntryType_EntryConfChange, Term: 1, Index: 1, Data: ccdata},
-			},
-			MustSync: true,
-		},
-		{
-			HardState:        pb.HardState{Term: 2, Commit: 3, Vote: 1},
-			Entries:          []pb.Entry{{Term: 2, Index: 3, Data: []byte("foo")}},
-			CommittedEntries: []pb.Entry{{Term: 2, Index: 3, Data: []byte("foo")}},
-			MustSync:         true,
-		},
-	}
-	storage := NewMemoryStorage()
-	c := &Config{
-		ID:            1,
-		ElectionTick:  10,
-		HeartbeatTick: 1,
-		Storage:       storage,
-		MaxEntsSize:   noLimit,
-	}
-	n := StartNode(c, []Peer{{ID: 1}})
-	defer n.Stop()
-	g := <-n.Ready()
-	if !reflect.DeepEqual(g, wants[0]) {
-		t.Fatalf("#%d: g = %+v,\n             w   %+v", 1, g, wants[0])
-	} else {
-		storage.Append(g.Entries)
-		n.Advance()
-	}
-
-	n.Campaign(ctx)
-	rd := <-n.Ready()
-	storage.Append(rd.Entries)
-	n.Advance()
-
-	n.Propose(ctx, []byte("foo"))
-	if g2 := <-n.Ready(); !reflect.DeepEqual(g2, wants[1]) {
-		t.Errorf("#%d: g = %+v,\n             w   %+v", 2, g2, wants[1])
-	} else {
-		storage.Append(g2.Entries)
-		n.Advance()
-	}
-
-	select {
-	case rd := <-n.Ready():
-		t.Errorf("unexpected Ready: %+v", rd)
-	case <-time.After(time.Millisecond):
-	}
-}
-
-func TestNodeRestart(t *testing.T) {
-	entries := []pb.Entry{
-		{Term: 1, Index: 1},
-		{Term: 1, Index: 2, Data: []byte("foo")},
-	}
-	st := pb.HardState{Term: 1, Commit: 1}
-
-	want := Ready{
-		HardState: st,
-		// commit up to index commit index in st
-		CommittedEntries: entries[:st.Commit],
-		MustSync:         true,
-	}
-
-	storage := NewMemoryStorage()
-	storage.SetHardState(st)
-	storage.Append(entries)
-	c := &Config{
-		ID:            1,
-		ElectionTick:  10,
-		HeartbeatTick: 1,
-		Storage:       storage,
-		MaxEntsSize:   noLimit,
-	}
-	n := RestartNode(c)
-	defer n.Stop()
-	if g := <-n.Ready(); !reflect.DeepEqual(g, want) {
-		t.Errorf("g = %+v,\n             w   %+v", g, want)
-	}
-	n.Advance()
-
-	select {
-	case rd := <-n.Ready():
-		t.Errorf("unexpected Ready: %+v", rd)
-	case <-time.After(time.Millisecond):
-	}
-}
-
-func TestNodeRestartFromSnapshot(t *testing.T) {
-	snap := pb.Snapshot{
-		Metadata: &pb.SnapshotMetadata{
-			ConfState: &pb.ConfState{Nodes: []uint64{1, 2}},
-			Index:     2,
-			Term:      1,
-		},
-	}
-	entries := []pb.Entry{
-		{Term: 1, Index: 3, Data: []byte("foo")},
-	}
-	st := pb.HardState{Term: 1, Commit: 3}
-
-	want := Ready{
-		HardState: st,
-		// commit up to index commit index in st
-		CommittedEntries: entries,
-		MustSync:         true,
-	}
-
-	s := NewMemoryStorage()
-	s.SetHardState(st)
-	s.ApplySnapshot(snap)
-	s.Append(entries)
-	c := &Config{
-		ID:            1,
-		ElectionTick:  10,
-		HeartbeatTick: 1,
-		Storage:       s,
-		MaxEntsSize:   noLimit,
-	}
-	n := RestartNode(c)
-	defer n.Stop()
-	if g := <-n.Ready(); !reflect.DeepEqual(g, want) {
-		t.Errorf("g = %+v,\n             w   %+v", g, want)
-	} else {
-		n.Advance()
-	}
-
-	select {
-	case rd := <-n.Ready():
-		t.Errorf("unexpected Ready: %+v", rd)
-	case <-time.After(time.Millisecond):
-	}
-}
-
-func TestNodeAdvance(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	storage := NewMemoryStorage()
-	c := &Config{
-		ID:            1,
-		ElectionTick:  10,
-		HeartbeatTick: 1,
-		Storage:       storage,
-		MaxEntsSize:   noLimit,
-	}
-	n := StartNode(c, []Peer{{ID: 1}})
-	defer n.Stop()
-	rd := <-n.Ready()
-	storage.Append(rd.Entries)
-	n.Advance()
-
-	n.Campaign(ctx)
-	<-n.Ready()
-
-	n.Propose(ctx, []byte("foo"))
-	select {
-	case rd = <-n.Ready():
-		t.Fatalf("unexpected Ready before Advance: %+v", rd)
-	case <-time.After(time.Millisecond):
-	}
-	storage.Append(rd.Entries)
-	n.Advance()
-	select {
-	case <-n.Ready():
-	case <-time.After(100 * time.Millisecond):
-		t.Errorf("expect Ready after Advance, but there is no Ready available")
-	}
-}
-
-func TestSoftStateEqual(t *testing.T) {
-	tests := []struct {
-		st *SoftState
-		we bool
-	}{
-		{&SoftState{}, true},
-		{&SoftState{Lead: 1}, false},
-		{&SoftState{RaftState: StateLeader}, false},
-	}
-	for i, tt := range tests {
-		if g := tt.st.equal(&SoftState{}); g != tt.we {
-			t.Errorf("#%d, equal = %v, want %v", i, g, tt.we)
-		}
-	}
-}
-
-func TestIsHardStateEqual(t *testing.T) {
-	tests := []struct {
-		st pb.HardState
-		we bool
-	}{
-		{emptyState, true},
-		{pb.HardState{Vote: 1}, false},
-		{pb.HardState{Commit: 1}, false},
-		{pb.HardState{Term: 1}, false},
-	}
-
-	for i, tt := range tests {
-		if isHardStateEqual(tt.st, emptyState) != tt.we {
-			t.Errorf("#%d, equal = %v, want %v", i, isHardStateEqual(tt.st, emptyState), tt.we)
-		}
-	}
-}
-
-func TestAppendPagination(t *testing.T) {
-	const maxEntsSize = 2048
-	n := newNetworkWithConfig(func(c *Config) {
-		c.MaxEntsSize = maxEntsSize
-	}, nil, nil, nil)
-
-	seenFullMessage := false
-	// Inspect all messages to see that we never exceed the limit, but
-	// we do see messages of larger than half the limit.
-	n.msgHook = func(m pb.Message) bool {
-		if m.MsgType == pb.MessageType_MsgAppend {
-			size := 0
-			for _, e := range m.Entries {
-				size += len(e.Data)
-			}
-			if size > maxEntsSize {
-				t.Errorf("sent MessageType_MsgAppend that is too large: %d bytes", size)
-			}
-			if size > maxEntsSize/2 {
-				seenFullMessage = true
-			}
-		}
-		return true
-	}
-
-	n.send(pb.Message{From: 1, To: 1, MsgType: pb.MessageType_MsgHup})
-
-	// Partition the network while we make our proposals. This forces
-	// the entries to be batched into larger messages.
-	n.isolate(1)
-	blob := []byte(strings.Repeat("a", 1000))
-	for i := 0; i < 5; i++ {
-		n.send(pb.Message{From: 1, To: 1, MsgType: pb.MessageType_MsgPropose, Entries: []*pb.Entry{{Data: blob}}})
-	}
-	n.recover()
-
-	// After the partition recovers, tick the clock to wake everything
-	// back up and send the messages.
-	n.send(pb.Message{From: 1, To: 1, MsgType: pb.MessageType_MsgBeat})
-	if !seenFullMessage {
-		t.Error("didn't see any messages more than half the max size; something is wrong with this test")
-	}
-}
-
-func TestCommitPagination(t *testing.T) {
-	s := NewMemoryStorage()
-	cfg := newTestConfig(1, []uint64{1}, 10, 1, s)
-	cfg.MaxEntsSize = 2048
-	r := newRaft(cfg)
-	n := newNode()
-	go n.run(r)
-	n.Campaign(context.TODO())
-
-	rd := readyWithTimeout(&n)
-	if len(rd.CommittedEntries) != 1 {
-		t.Fatalf("expected 1 (empty) entry, got %d", len(rd.CommittedEntries))
-	}
-	s.Append(rd.Entries)
-	n.Advance()
-
-	blob := []byte(strings.Repeat("a", 1000))
-	for i := 0; i < 3; i++ {
-		if err := n.Propose(context.TODO(), blob); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// The 3 proposals will commit in two batches.
-	rd = readyWithTimeout(&n)
-	if len(rd.CommittedEntries) != 2 {
-		t.Fatalf("expected 2 entries in first batch, got %d", len(rd.CommittedEntries))
-	}
-	s.Append(rd.Entries)
-	n.Advance()
-	rd = readyWithTimeout(&n)
-	if len(rd.CommittedEntries) != 1 {
-		t.Fatalf("expected 1 entry in second batch, got %d", len(rd.CommittedEntries))
-	}
-	s.Append(rd.Entries)
-	n.Advance()
-}
-
-type ignoreSizeHintMemStorage struct {
-	*MemoryStorage
-}
-
-func (s *ignoreSizeHintMemStorage) Entries(lo, hi uint64, maxSize uint64) ([]pb.Entry, error) {
-	return s.MemoryStorage.Entries(lo, hi, math.MaxUint64)
-}
-
-// TestNodeCommitPaginationAfterRestart regression tests a scenario in which the
-// Storage's Entries size limitation is slightly more permissive than Raft's
-// internal one. The original bug was the following:
-//
-// - node learns that index 11 (or 100, doesn't matter) is committed
-// - nextEnts returns index 1..10 in CommittedEntries due to size limiting. However,
-//   index 10 already exceeds maxBytes, due to a user-provided impl of Entries.
-// - Commit index gets bumped to 10
-// - the node persists the HardState, but crashes before applying the entries
-// - upon restart, the storage returns the same entries, but `slice` takes a different code path
-//   (since it is now called with an upper bound of 10) and removes the last entry.
-// - Raft emits a HardState with a regressing commit index.
-//
-// A simpler version of this test would have the storage return a lot less entries than dictated
-// by maxSize (for example, exactly one entry) after the restart, resulting in a larger regression.
-// This wouldn't need to exploit anything about Raft-internal code paths to fail.
-func TestNodeCommitPaginationAfterRestart(t *testing.T) {
-	s := &ignoreSizeHintMemStorage{
-		MemoryStorage: NewMemoryStorage(),
-	}
-	persistedHardState := pb.HardState{
-		Term:   1,
-		Vote:   1,
-		Commit: 10,
-	}
-
-	s.hardState = persistedHardState
-	s.ents = make([]pb.Entry, 10)
-	var size uint64
-	for i := range s.ents {
-		ent := pb.Entry{
-			Term:      1,
-			Index:     uint64(i + 1),
-			EntryType: pb.EntryType_EntryNormal,
-			Data:      []byte("a"),
-		}
-
-		s.ents[i] = ent
-		size += uint64(ent.Size())
-	}
-
-	cfg := newTestConfig(1, []uint64{1}, 10, 1, s)
-	// Set a MaxEntsSize that would suggest to Raft that the last committed entry should
-	// not be included in the initial rd.CommittedEntries. However, our storage will ignore
-	// this and *will* return it (which is how the Commit index ended up being 10 initially).
-	cfg.MaxEntsSize = size - uint64(s.ents[len(s.ents)-1].Size()) - 1
-
-	r := newRaft(cfg)
-	n := newNode()
-	go n.run(r)
-	defer n.Stop()
-
-	rd := readyWithTimeout(&n)
-	if !IsEmptyHardState(rd.HardState) && rd.HardState.Commit < persistedHardState.Commit {
-		t.Errorf("HardState regressed: Commit %d -> %d\nCommitting:\n%+v",
-			persistedHardState.Commit, rd.HardState.Commit,
-			DescribeEntries(rd.CommittedEntries, func(data []byte) string { return fmt.Sprintf("%q", data) }),
-		)
-	}
-}
